Pulls most functions out of Status and into StatusUtils. Fixes run duration bug
