+++ frontend/src/pages/Status.test.tsx
@@ -15,7 +15,8 @@
  */
 
 import * as Utils from '../lib/Utils';
-import { NodePhase, hasFinished, statusBgColors, statusToBgColor, statusToIcon, checkIfTerminated } from './Status';
+import { statusToIcon } from './Status';
+import { NodePhase } from '../lib/StatusUtils';
 import { shallow } from 'enzyme';
 
 
@@ -74,105 +75,5 @@ describe('Status', () => {
         expect(tree).toMatchSnapshot();
       })
     ));
-  });
-
-  describe('hasFinished', () => {
-    [NodePhase.ERROR, NodePhase.FAILED, NodePhase.SUCCEEDED, NodePhase.SKIPPED, NodePhase.TERMINATED].forEach(status => {
-      it(`returns \'true\' if status is: ${status}`, () => {
-        expect(hasFinished(status)).toBe(true);
-      });
-    });
-
-    [NodePhase.PENDING, NodePhase.RUNNING, NodePhase.UNKNOWN, NodePhase.TERMINATING].forEach(status => {
-      it(`returns \'false\' if status is: ${status}`, () => {
-        expect(hasFinished(status)).toBe(false);
-      });
-    });
-
-    it('returns \'false\' if status is undefined', () => {
-      expect(hasFinished(undefined)).toBe(false);
-    });
-
-    it('returns \'false\' if status is invalid', () => {
-      expect(hasFinished('bad phase' as any)).toBe(false);
-    });
-  });
-
-  describe('statusToBgColor', () => {
-    it('handles an invalid phase', () => {
-      const consoleSpy = jest.spyOn(console, 'log').mockImplementationOnce(() => null);
-      expect(statusToBgColor('bad phase' as any)).toEqual(statusBgColors.notStarted);
-      expect(consoleSpy).toHaveBeenLastCalledWith('Unknown node phase:', 'bad phase');
-    });
-
-    it('handles an \'Unknown\' phase', () => {
-      const consoleSpy = jest.spyOn(console, 'log').mockImplementationOnce(() => null);
-      expect(statusToBgColor(NodePhase.UNKNOWN)).toEqual(statusBgColors.notStarted);
-      expect(consoleSpy).toHaveBeenLastCalledWith('Unknown node phase:', 'Unknown');
-    });
-
-    it('returns color \'not started\' if status is undefined', () => {
-      const consoleSpy = jest.spyOn(console, 'log').mockImplementationOnce(() => null);
-      expect(statusToBgColor(undefined)).toEqual(statusBgColors.notStarted);
-      expect(consoleSpy).toHaveBeenLastCalledWith('Unknown node phase:', undefined);
-    });
-
-    it('returns color \'not started\' if status is \'Pending\'', () => {
-      expect(statusToBgColor(NodePhase.PENDING)).toEqual(statusBgColors.notStarted);
-    });
-
-    [NodePhase.ERROR, NodePhase.FAILED].forEach(status => {
-      it(`returns color \'error\' if status is: ${status}`, () => {
-        expect(statusToBgColor(status)).toEqual(statusBgColors.error);
-      });
-    });
-
-    [NodePhase.RUNNING, NodePhase.TERMINATING].forEach(status => {
-      it(`returns color \'running\' if status is: ${status}`, () => {
-        expect(statusToBgColor(status)).toEqual(statusBgColors.running);
-      });
-    });
-
-    [NodePhase.SKIPPED, NodePhase.TERMINATED].forEach(status => {
-      it(`returns color \'terminated or skipped\' if status is: ${status}`, () => {
-        expect(statusToBgColor(status)).toEqual(statusBgColors.terminatedOrSkipped);
-      });
-    });
-
-    it('returns color \'succeeded\' if status is \'Succeeded\'', () => {
-      expect(statusToBgColor(NodePhase.SUCCEEDED)).toEqual(statusBgColors.succeeded);
-    });
-  });
-
-  describe('checkIfTerminated', () => {
-    it('returns status \'terminated\' if status is \'failed\' and error message is \'terminated\'', () => {
-      expect(checkIfTerminated(NodePhase.FAILED, 'terminated')).toEqual(NodePhase.TERMINATED);
-    });
-
-    [
-      NodePhase.SUCCEEDED,
-      NodePhase.ERROR,
-      NodePhase.SKIPPED,
-      NodePhase.PENDING,
-      NodePhase.RUNNING,
-      NodePhase.TERMINATING,
-      NodePhase.UNKNOWN
-    ].forEach(status => {
-      it(`returns the original status, even if message is 'terminated', if status is: ${status}`, () => {
-        expect(checkIfTerminated(status, 'terminated')).toEqual(status);
-      });
-    });
-
-    it('returns \'failed\' if status is \'failed\' and no error message is provided', () => {
-      expect(checkIfTerminated(NodePhase.FAILED)).toEqual(NodePhase.FAILED);
-    });
-
-    it('returns \'failed\' if status is \'failed\' and empty error message is provided', () => {
-      expect(checkIfTerminated(NodePhase.FAILED, '')).toEqual(NodePhase.FAILED);
-    });
-
-    it('returns \'failed\' if status is \'failed\' and arbitrary error message is provided', () => {
-      expect(checkIfTerminated(NodePhase.FAILED, 'some random error')).toEqual(NodePhase.FAILED);
-    });
   });
 });
