+++ frontend/src/pages/Status.tsx
@@ -25,81 +25,7 @@ import Tooltip from '@material-ui/core/Tooltip';
 import UnknownIcon from '@material-ui/icons/Help';
 import { color } from '../Css';
 import { logger, formatDateString } from '../lib/Utils';
-
-export const statusBgColors = {
-  error: '#fce8e6',
-  notStarted: '#f7f7f7',
-  running: '#e8f0fe',
-  succeeded: '#e6f4ea',
-  terminatedOrSkipped: '#f1f3f4',
-  warning: '#fef7f0',
-};
-
-export enum NodePhase {
-  ERROR = 'Error',
-  FAILED = 'Failed',
-  PENDING = 'Pending',
-  RUNNING = 'Running',
-  SKIPPED = 'Skipped',
-  SUCCEEDED = 'Succeeded',
-  TERMINATING = 'Terminating',
-  TERMINATED = 'Terminated',
-  UNKNOWN = 'Unknown',
-}
-
-export function hasFinished(status?: NodePhase): boolean {
-  switch (status) {
-    case NodePhase.SUCCEEDED: // Fall through
-    case NodePhase.FAILED: // Fall through
-    case NodePhase.ERROR: // Fall through
-    case NodePhase.SKIPPED: // Fall through
-    case NodePhase.TERMINATED:
-      return true;
-    case NodePhase.PENDING: // Fall through
-    case NodePhase.RUNNING: // Fall through
-    case NodePhase.TERMINATING: // Fall through
-    case NodePhase.UNKNOWN:
-      return false;
-    default:
-      return false;
-  }
-}
-
-export function statusToBgColor(status?: NodePhase, nodeMessage?: string): string {
-  status = checkIfTerminated(status, nodeMessage);
-  switch (status) {
-    case NodePhase.ERROR:
-      // fall through
-    case NodePhase.FAILED:
-      return statusBgColors.error;
-    case NodePhase.PENDING:
-      return statusBgColors.notStarted;
-    case NodePhase.TERMINATING:
-      // fall through
-    case NodePhase.RUNNING:
-      return statusBgColors.running;
-    case NodePhase.SUCCEEDED:
-      return statusBgColors.succeeded;
-    case NodePhase.SKIPPED:
-      // fall through
-    case NodePhase.TERMINATED:
-      return statusBgColors.terminatedOrSkipped;
-    case NodePhase.UNKNOWN:
-      // fall through
-    default:
-      logger.verbose('Unknown node phase:', status);
-      return statusBgColors.notStarted;
-  }
-}
-
-export function checkIfTerminated(status?: NodePhase, nodeMessage?: string): NodePhase | undefined {
-  // Argo considers terminated runs as having "Failed", so we have to examine the failure message to
-  // determine why the run failed.
-  if (status === NodePhase.FAILED && nodeMessage === 'terminated') {
-    status = NodePhase.TERMINATED;
-  }
-  return status;
-}
+import { NodePhase, checkIfTerminated } from '../lib/StatusUtils';
 
 export function statusToIcon(status?: NodePhase, startDate?: Date | string, endDate?: Date | string, nodeMessage?: string): JSX.Element {
   status = checkIfTerminated(status, nodeMessage);
@@ -152,7 +78,6 @@ export function statusToIcon(status?: NodePhase, startDate?: Date | string, endD
     default:
       logger.verbose('Unknown node phase:', status);
   }
-
   return (
     <Tooltip title={
       <div>
